[Feature] Add about screen (T15154)
In this merge request, we have implemented the about screen to show useful links, authors, and third-party libraries as well as an abstract view to display any ```NSAttributedString``` for compatibility with iOS 14.

As we do not sort deserialized arrays in code, future maintainers of ```contributors.json``` and ```libs.json``` should keep the order by ```name``` for ```contributors.json``` and ```id``` (project name) for ```libs.json```.

There are two places that can be improved by future development:
1. for function ```func getContributorListAttributedText(template: String, for category: KeyPath<Contributors, [Contributor]>) -> NSAttributedString```, the complexity of the current implementation is $`O(n^2)`$, and it might be optimized to $`O(n)`$
2. for section ```Actions```, we have tried ```LazyHGrid```, but it is not flexible enough for enlarged text (accessibility) and landscape orientation, so future development should take a look on how to implement a list similar to the Android app

Screenshots:
![image](/uploads/d462ff4c1a3b215030b2bf877bfadb10/image.png)
![image](/uploads/b6169a2d16c323f54e5921bc434336c8/image.png)