Change certificate validity period to match other platforms
When a certificate has expired, any connection attempt just fails with no 
indication in the GUI, see e.g. https://github.com/NixOS/nixpkgs/issues/240606.
Ideally the GUI of all clients would be improved to handle this in a better way,
but the problem especially affected iOS devices, as the Android and desktop
clients generate certificates with 10 years validity instead of just 1 on iOS.